elf_aux_info(3): accept a "big enough" buffer
In some cases, elf_aux_info(3)'s buflen parameter is checked to be
exactly the size of the entry's underlying type. This can be difficult
to use since the sizes required vary per-entry and are not documented.

Fix this by accepting big enough buffers for all entries.

Note that the entries AT_CANARY and AT_PAGESIZES originally accepted a
buffer smaller than or equal to their length, since there was no way for
a user to know the amount of valid data returned. However, now it is
acceptable to copy to a larger buffer since the user can query
AT_CANARYLEN and AT_PAGESIZESLEN.