elf_aux_info(3): expose all auxv entries
elf_aux_info(3) is a convenient wrapper for userspace access to the ELF
auxiliary vector. However, only a subset of the entries are exposed
through this interface. Update it to include all of the currently
defined entries other than AT_NOTELF.

Note that some of these entries (such as AT_UID, AT_GID, etc) are
only exported for the Linux ABI, and are marked as such.

PR:     234523